Move method to ServerAmpleImpl
